:sparkles: feat: configure routes
